MXS-4791 Change all uses of size_t and ssize_t to int64_t
Uncrustify all wcar code.

+++ server/modules/filter/wcar/player/wcarplayerconfig.cc
@@ -31,7 +31,7 @@ constexpr int INDENT = 12;
 template<typename H>
 auto OPT(int optval, H help)
 {
-    size_t idx;
+    int64_t idx;
     for (idx = 0; long_opts[idx].val; ++idx)
     {
         if (long_opts[idx].val == optval)
